Breaks out enrollment flow into its own component. Adds profile landing page with stubbed-out buttons for navigating to enrollment info forms.
