+++ frontend/src/app/enrollment-flow/enrollment-flow.component.html
@@ -1,0 +1,69 @@
+<div
+  *ngIf="user"
+  class="profile mat-typography"
+>
+  <h1>{{user.first_name}} {{user.last_name}}’s enrollment info</h1>
+  <div>
+    <mat-progress-bar
+      mode="determinate"
+      [value]="10"
+    ></mat-progress-bar>
+
+
+    <app-loading
+      *ngIf="loading"
+      message="Loading..."
+    ></app-loading>
+
+    <mat-drawer-container *ngIf="!loading">
+      <mat-drawer
+        *ngIf="user && stepNames"
+        [mode]="'side'"
+        [opened]="true"
+      >
+        <app-questionnaire-steps-list
+          [user]="user"
+          [stepNames]="stepNames"
+        ></app-questionnaire-steps-list>
+      </mat-drawer>
+      <mat-drawer-content>
+        <form
+          *ngIf="form && step && options"
+          [formGroup]="form"
+          (ngSubmit)="submit()"
+          class="pad-2"
+        >
+          <h1>{{ step.label }}</h1>
+          <p>{{ step.description }}</p>
+
+          <formly-form
+            [form]="form.at(activeStep)"
+            [model]="model"
+            [fields]="step.fields"
+            [options]="options[activeStep]"
+          >
+          </formly-form>
+
+          <div class="button-row">
+            <button
+              mat-flat-button
+              type="button"
+              (click)="prevStep(activeStep)"
+            >Back</button>
+            <button
+              mat-flat-button
+              type="button"
+              [disabled]="!form.at(activeStep).valid"
+              (click)="nextStep(activeStep)"
+            >Save &amp; Next</button>
+            <button
+              mat-flat-button
+              [disabled]="!form.valid"
+              type="submit"
+            >Submit</button>
+          </div>
+        </form>
+      </mat-drawer-content>
+    </mat-drawer-container>
+  </div>
+</div>
