+++ frontend/src/app/participant-profile/participant-profile.component.scss
@@ -1,0 +1,18 @@
+[mat-card-avatar] {
+  width: 100px;
+  height: 100px;
+}
+
+mat-card-subtitle {
+  mat-progress-bar {
+    height: 16px;
+  }
+
+  button {
+    margin-top: 20px;
+  }
+
+  .mat-subheading-2 {
+    color: black;
+  }
+}
