+++ frontend/src/app/participant.ts
@@ -1,0 +1,11 @@
+export interface Participant {
+
+  first_name: string;
+  nickname: string;
+  last_name: string;
+  relationship: string;
+  percent_complete: number;
+  num_studies_enrolled: number;
+  avatar_img_url: string;
+
+}
