+++ frontend/src/app/profile/profile.component.ts
@@ -1,12 +1,7 @@
 import { Component, OnInit } from '@angular/core';
-import { FormArray, FormGroup } from '@angular/forms';
-import { ActivatedRoute, Router } from '@angular/router';
-import { FormlyFormOptions } from '@ngx-formly/core';
-import * as flatten from 'flat';
-import { keysToCamel, toCamel } from 'src/util/snakeToCamel';
 import { ApiService } from '../api.service';
-import { QuestionnaireStep } from '../step';
 import { User } from '../user';
+import { Participant } from '../participant';
 
 
 @Component({
@@ -19,156 +14,43 @@ export class ProfileComponent implements OnInit {
   stepName: string;
   activeStep = 0;
   loading = true;
-  stepNames = ['identification', 'contact', 'demographics', 'home', 'evaluation_history'];
-
-  model = {};
-  step: QuestionnaireStep;
-  form: FormArray;
-  options;
+  participants: Participant[] = [
+    {
+      first_name: 'Firstname',
+      nickname: '',
+      last_name: 'Lastname',
+      relationship: 'self',
+      percent_complete: 20,
+      num_studies_enrolled: 0,
+      avatar_img_url: 'https://pbs.twimg.com/profile_images/477556215401025537/zH_q0-_s.png'
+    },
+    {
+      first_name: 'Childname',
+      nickname: 'Nickname',
+      last_name: 'Lastname',
+      relationship: 'dependent',
+      percent_complete: 100,
+      num_studies_enrolled: 3,
+      avatar_img_url: 'https://pbs.twimg.com/profile_images/493362921817120770/23Hg2d2k.jpeg'
+    },
+  ];
 
   constructor(
-    private api: ApiService,
-    private router: Router,
-    private route: ActivatedRoute
+    private api: ApiService
   ) {
-    console.log('this.stepNames', this.stepNames);
 
     this.api.getSession().subscribe(user => {
       this.user = user;
-
-      console.log('user', user);
-
-      this.route.params.subscribe(params => {
-        const stepName = params.stepName || '';
-
-        if (stepName !== '') {
-          this.api.getQuestionnaireMeta(stepName).subscribe(q => {
-            this.step = this.infoToFormlyForm(q.get_meta, stepName);
-            this.form = new FormArray([new FormGroup({})]);
-            this.options = <FormlyFormOptions>{};
-            this.loading = false;
-          });
-
-        } else {
-          this.loading = false;
-        }
-      });
+      this.loading = false;
     }, error1 => {
       this.user = null;
+      this.loading = false;
     });
   }
 
   ngOnInit() {
   }
 
-  infoToFormlyForm(info, stepName): QuestionnaireStep {
-    const step = new QuestionnaireStep({
-      id: info.id,
-      name: stepName,
-      label: '',
-      description: '',
-      fields: []
-    });
-    const stepFields = [];
 
-    Object.keys(info).forEach(key => {
-      const item = info[key];
-
-      if (key === 'table') {
-        Object.keys(item).forEach(tableKey => {
-          step[toCamel(tableKey)] = item[tableKey];
-        });
-      } else if (key === 'field_groups') {
-        Object.keys(item).forEach(wrapperKey => {
-
-          // Clone the wrapper object so we can delete the original later
-          const wrapper = this.clone(item[wrapperKey]);
-          const fgFields = item[wrapperKey].fields || [];
-          wrapper.key = wrapperKey;
-          wrapper.fieldGroup = fgFields.map((childKey: string) => {
-            const childField = this.clone(info[childKey]);
-            childField.key = childKey;
-            childField.name = childKey;
-
-            // Remove the field from the 'all' object, since we've
-            // now copied it to its parent fieldGroup
-            delete info[childKey];
-            return keysToCamel(childField);
-          });
-
-          // Remove the fields array from the wrapper object,
-          // since all its child fields are now inside the
-          // fieldGroup attribute
-          delete wrapper.fields;
-          stepFields.push(keysToCamel(wrapper));
-        });
-      } else if (item) {
-        item.key = key;
-        item.name = key;
-        stepFields.push(keysToCamel(item));
-      }
-    });
-
-    stepFields.sort((f1, f2) => f1.displayOrder - f2.displayOrder);
-    stepFields.forEach(f => step.fields.push(f));
-
-    console.log('step', step);
-
-    return step;
-  }
-
-  prevStep(step: number) {
-    this.activeStep = step - 1;
-    this.submit();
-  }
-
-  nextStep(step: number) {
-    this.activeStep = step + 1;
-    this.submit();
-  }
-
-  setActiveStep(step: number) {
-    this.activeStep = step;
-    this.submit();
-  }
-
-  submit() {
-    console.log('this.model', this.model);
-
-    // Flatten the model
-    const flattened = flatten(this.model, { safe: true });
-
-    console.log('flattened', flattened);
-
-
-    // Rename the keys
-    const options = {};
-    const pattern = /^(.*)\./gi;
-    Object.keys(flattened).map(oldKey => {
-      const newKey = oldKey.replace(pattern, '');
-      options[newKey] = flattened[oldKey];
-    });
-
-    if (isFinite(this.step.id)) {
-      this.api.updateQuestionnaire(this.step.name, this.step.id, options).subscribe(response => {
-        // Update form with saved values
-        this.router.navigate(['profile', this.stepNames[this.activeStep]]);
-      });
-    } else {
-      console.log('this.step.name', this.step.name);
-
-
-      this.api.submitQuestionnaire(this.step.name, options).subscribe(response => {
-        // Update form with saved values
-        console.log('this.stepNames[this.activeStep]', this.stepNames[this.activeStep]);
-
-        this.router.navigate(['profile', this.stepNames[this.activeStep]]);
-      });
-    }
-  }
-
-  clone(o: any): any {
-    return JSON.parse(JSON.stringify(o));
-  }
 
 }
