+++ frontend/src/app/questionnaire-steps-list/questionnaire-steps-list.component.html
@@ -2,7 +2,7 @@
   <a
     *ngIf="step && step.name"
     mat-list-item
-    [routerLink]="['/profile/' + step.name]"
+    [routerLink]="['/profile/enrollment/' + step.name]"
     routerLinkActive="router-link-active"
   >{{step.label}}</a>
 </mat-nav-list>
