+++ frontend/src/app/routing/routing.module.ts
@@ -14,6 +14,7 @@ import { ResourceDetailComponent } from '../resource-detail/resource-detail.comp
 import { StudyDetailComponent } from '../study-detail/study-detail.component';
 import { TrainingDetailComponent } from '../training-detail/training-detail.component';
 import { QuestionnaireStepComponent } from '../questionnaire-step/questionnaire-step.component';
+import { EnrollmentFlowComponent } from '../enrollment-flow/enrollment-flow.component';
 
 const routes: Routes = [
   { path: '', redirectTo: 'home', pathMatch: 'full' },
@@ -26,7 +27,8 @@ const routes: Routes = [
     data: { title: 'Reset your STAR Drive password', hideHeader: true }
   },
   { path: 'profile', component: ProfileComponent, data: { title: 'Your STAR Drive Account' } },
-  { path: 'profile/:stepName', component: ProfileComponent, data: { title: 'Your STAR Drive Account' } },
+  { path: 'profile/enrollment', component: EnrollmentFlowComponent, data: { title: 'Your STAR Drive Account' } },
+  { path: 'profile/enrollment/:stepName', component: EnrollmentFlowComponent, data: { title: 'Your STAR Drive Account' } },
   { path: 'register', component: RegisterComponent, data: { title: 'Create a STAR Drive Account', hideHeader: true } },
   { path: 'resources', component: ResourcesComponent, data: { title: 'View STAR Drive Trainings & Resources' } },
   { path: 'resource/:resourceId', component: ResourceDetailComponent, data: { title: 'Resource Details' } },
